--- blog-project/src/app.module.ts
@@ -1,15 +1,0 @@
-import { Module } from '@nestjs/common';
-import { TypeOrmModule } from '@nestjs/typeorm';
-import { BoardsModule } from './boards/boards.module';
-import { typeORMConfig } from './configs/typrorm.config';
-import { AuthModule } from './auth/auth.module';
-
-@Module({
-  imports: [
-    TypeOrmModule.forRoot(typeORMConfig),
-    BoardsModule,
-    AuthModule
-  ],
-})
-
-export class AppModule {}
