--- blog-project/src/auth/auth.module.ts
@@ -1,27 +1,0 @@
-import { Module } from '@nestjs/common';
-import { JwtModule } from '@nestjs/jwt';
-import { PassportModule } from '@nestjs/passport';
-import { TypeOrmModule } from '@nestjs/typeorm';
-import { JwtStrategy } from './jwt.strategy';
-import { AuthController } from './auth.controller';
-import { AuthService } from './auth.service';
-import { UserRepository } from './user.repository';
-import * as config from 'config';
-
-const jwtConfig = config.get('jwt');
-@Module({
-  imports: [ 
-    PassportModule.register({ defaultStrategy: 'jwt'}),
-    JwtModule.register({
-      secret: process.env.JWT_SECRET || jwtConfig.secret,
-      signOptions: {
-        expiresIn: jwtConfig.expiresIn
-      }
-    }),
-    TypeOrmModule.forFeature([UserRepository])
-  ],
-  controllers: [AuthController],
-  providers: [AuthService, JwtStrategy],
-  exports: [JwtStrategy, PassportModule]
-})
-export class AuthModule {}
