--- blog-project/src/auth/jwt.strategy.ts
@@ -1,31 +1,0 @@
-import { Injectable, UnauthorizedException } from "@nestjs/common";
-import { PassportStrategy } from "@nestjs/passport";
-import { InjectRepository } from "@nestjs/typeorm";
-import { ExtractJwt, Strategy } from "passport-jwt";
-import { User } from "./user.entity";
-import { UserRepository } from "./user.repository";
-import * as config from 'config';
-
-@Injectable()
-export class JwtStrategy extends PassportStrategy(Strategy) {
-    constructor(
-        @InjectRepository(UserRepository)
-        private userRepository: UserRepository
-    ) {
-        super({
-            secretOrKey: process.env.JWT_SECRET || config.get('jwt.secret'),
-            jwtFromRequest: ExtractJwt.fromAuthHeaderAsBearerToken()
-        })
-    }
-
-    async validate(payload) {
-        const { username } = payload;
-        const user: User = await this.userRepository.findOne({ username });
-
-        if(!user) {
-            throw new UnauthorizedException();
-        }
-
-        return user;
-    }
-}
