--- blog-project/src/boards/board-status.enum.ts
@@ -1,4 +1,0 @@
-export enum BoardStatus {
-    PUBLIC = 'PUBLIC',
-    PRIVATE = 'PRIVATE'
-}
