--- blog-project/src/boards/board.entity.ts
@@ -1,21 +1,0 @@
-import { User } from "src/auth/user.entity";
-import { BaseEntity, Column, Entity, ManyToOne, PrimaryGeneratedColumn } from "typeorm";
-import { BoardStatus } from "./board-status.enum";
-
-@Entity()
-export class Board extends BaseEntity {
-    @PrimaryGeneratedColumn()
-    id: number;
-
-    @Column()
-    title: string;
-
-    @Column()
-    description: string;
-
-    @Column()
-    status: BoardStatus;
-
-    @ManyToOne(type => User, user => user.boards, { eager: false })
-    user: User;
-}
