--- blog-project/src/boards/boards.controller.ts
@@ -1,54 +1,0 @@
-import { Body, Controller, Delete, Get, Logger, Param, ParseIntPipe, Patch, Post, UseGuards, UsePipes, ValidationPipe } from '@nestjs/common';
-import { AuthGuard } from '@nestjs/passport';
-import { GetUser } from 'src/auth/get-user.decorator';
-import { User } from 'src/auth/user.entity';
-import { BoardStatus } from './board-status.enum';
-import { Board } from './board.entity';
-import { BoardsService } from './boards.service';
-import { CreateBoardDto } from './dto/create-board.dto';
-import { BoardStatusValidationPipe } from './pipes/board-status-validation.pipe';
-
-@Controller('boards')
-@UseGuards(AuthGuard())
-export class BoardsController {
-    private logger = new Logger('Boards');
-
-    constructor(private boardsService: BoardsService) { }
-    
-    @Get()
-    getAllBoard(
-        @GetUser() user: User
-        ): Promise<Board[]> {
-            this.logger.verbose(`User ${user.username} trying to get all board`);
-            return this.boardsService.getAllBoards(user);
-    }
-
-    @Get('/:id')
-    getBoardById(@Param('id') id: number): Promise<Board> {
-        return this.boardsService.getBoardById(id);
-    }
-
-    @Post()
-    @UsePipes(ValidationPipe)
-    createBoard(@Body() createBoardDto: CreateBoardDto,
-    @GetUser() user: User): Promise<Board> {
-        this.logger.verbose(`User ${user.username} created a new board.
-        Payload: ${JSON.stringify(createBoardDto)}`);
-        return this.boardsService.createBoard(createBoardDto, user);
-    }
-
-    @Delete('/:id')
-    deleteBoard(@Param('id', ParseIntPipe) id: number,
-    @GetUser() user: User
-    ): Promise<void> {
-        return this.boardsService.deleteBoard(id, user);
-    }
-
-    @Patch('/:id/status')
-    updateBoardStatus(
-        @Param('id', ParseIntPipe) id: number,
-        @Body('status', BoardStatusValidationPipe) status: BoardStatus,
-    ): Promise<Board> {
-        return this.boardsService.updateBoardStatus(id, status);
-    }
-}
