--- blog-project/src/boards/boards.module.ts
@@ -1,16 +1,0 @@
-import { Module } from '@nestjs/common';
-import { TypeOrmModule } from '@nestjs/typeorm';
-import { AuthModule } from 'src/auth/auth.module';
-import { BoardRepository } from './board.repository';
-import { BoardsController } from './boards.controller';
-import { BoardsService } from './boards.service';
-
-@Module({
-  imports: [
-    TypeOrmModule.forFeature([BoardRepository]),
-    AuthModule
-  ],
-  controllers: [BoardsController],
-  providers: [BoardsService]
-})
-export class BoardsModule {}
