--- blog-project/src/boards/dto/create-board.dto.ts
@@ -1,9 +1,0 @@
-import { IsNotEmpty } from "class-validator";
-
-export class CreateBoardDto {
-    @IsNotEmpty()
-    title: string;
-
-    @IsNotEmpty()
-    description: string;
-}
