+++ src/app/app.module.ts
@@ -1,0 +1,10 @@
+import { Module } from '@nestjs/common';
+import { TypeOrmModule } from '@nestjs/typeorm';
+import { BoardsModule } from '../modules/boards.module';
+import { typeORMConfig } from '../configs/typrorm.config';
+import { AuthModule } from '../modules/user.module';
+
+@Module({
+  imports: [TypeOrmModule.forRoot(typeORMConfig), BoardsModule, AuthModule],
+})
+export class AppModule {}
