+++ src/configs/typrorm.config.ts
@@ -1,0 +1,15 @@
+import { TypeOrmModuleOptions } from '@nestjs/typeorm';
+import * as config from 'config';
+
+const dbConfig = config.get('db');
+
+export const typeORMConfig: TypeOrmModuleOptions = {
+  type: process.env.DB_TYPE || dbConfig.type,
+  host: process.env.DB_HOSTNAME || dbConfig.host,
+  port: process.env.DB_PORT || dbConfig.port,
+  username: process.env.DB_USERNAME || dbConfig.username,
+  password: process.env.DB_PASSWORD || dbConfig.password,
+  database: process.env.DB_NAME || dbConfig.database,
+  entities: [__dirname + '/../**/*.entity.{js,ts}'],
+  synchronize: process.env.DB_SYNCHRONIZE || dbConfig.synchronize,
+};
