+++ src/controllers/users.controller.ts
@@ -1,0 +1,29 @@
+import { Body, Controller, Post, UseGuards, ValidationPipe } from '@nestjs/common';
+import { AuthGuard } from '@nestjs/passport';
+import { AuthService } from '../services/user.service';
+import { AuthCredentialsDto } from '../dto/user-credential.dto';
+import { GetUser } from '../utils/get-user.decorator';
+import { User } from '../entities/user.entity';
+
+@Controller('auth')
+export class AuthController {
+  constructor(private authService: AuthService) {}
+
+  @Post('/signup')
+  signUp(@Body(ValidationPipe) authCredentialsDto: AuthCredentialsDto): Promise<void> {
+    return this.authService.signUp(authCredentialsDto);
+  }
+
+  @Post('/signin')
+  signIn(
+    @Body(ValidationPipe) authCredentialsDto: AuthCredentialsDto,
+  ): Promise<{ accessToken: string }> {
+    return this.authService.signIn(authCredentialsDto);
+  }
+
+  @Post('/test1')
+  @UseGuards(AuthGuard())
+  test(@GetUser() user: User) {
+    console.log('user', user);
+  }
+}
