+++ src/modules/user.module.ts
@@ -1,0 +1,27 @@
+import { Module } from '@nestjs/common';
+import { JwtModule } from '@nestjs/jwt';
+import { PassportModule } from '@nestjs/passport';
+import { TypeOrmModule } from '@nestjs/typeorm';
+import * as config from 'config';
+import { JwtStrategy } from '../jwt/jwt.strategy';
+import { AuthController } from '../controllers/users.controller';
+import { AuthService } from '../services/user.service';
+import { UserRepository } from '../repositories/user.repository';
+
+const jwtConfig = config.get('jwt');
+@Module({
+  imports: [
+    PassportModule.register({ defaultStrategy: 'jwt' }),
+    JwtModule.register({
+      secret: process.env.JWT_SECRET || jwtConfig.secret,
+      signOptions: {
+        expiresIn: process.env.EXPIRES_IN || jwtConfig.expiresIn,
+      },
+    }),
+    TypeOrmModule.forFeature([UserRepository]),
+  ],
+  controllers: [AuthController],
+  providers: [AuthService, JwtStrategy],
+  exports: [JwtStrategy, PassportModule],
+})
+export class AuthModule {}
