+++ src/pipes/board-status-validation.pipe.ts
@@ -1,0 +1,20 @@
+import { BadRequestException, PipeTransform } from '@nestjs/common';
+import { BoardStatus } from '../utils/board-status.enum';
+
+export class BoardStatusValidationPipe implements PipeTransform {
+  readonly StatusOptions = [BoardStatus.PRIVATE, BoardStatus.PUBLIC];
+
+  transform(value: any): any {
+    if (!this.isStatusValid(value)) {
+      throw new BadRequestException(`${value} is not a valid board status`);
+    }
+
+    return value;
+  }
+
+  private isStatusValid(status: any): boolean {
+    const index = this.StatusOptions.indexOf(status);
+
+    return index !== -1;
+  }
+}
