+++ src/utils/board-status.enum.ts
@@ -1,0 +1,4 @@
+export enum BoardStatus {
+  PUBLIC = 'PUBLIC',
+  PRIVATE = 'PRIVATE',
+}
